requirements.txt
Django==4.1
psycopg2
python-dotenv
django-money
djangorestframework
markdown
django-filter
python-dateutil